refactor(Policies): Remove GraphQL implementation
